+++ tests/features/terrain.py
@@ -1,5 +1,7 @@
 # terrain.py
 import os
+import shutil
+
 from lettuce import before, after, world
 
 from bigml.api import BigML
@@ -37,6 +39,10 @@ def setup_resources(feature):
 
 @after.each_feature
 def cleanup_resources(feature):
+
+    if os.path.exists('./tmp'):
+        shutil.rmtree('./tmp')
+
     for id in world.sources:
         world.api.delete_source(id)
     world.sources = []